+++ R/zzz.R
@@ -73,6 +73,7 @@ pkg.env$np         <- NULL
     # LOAD MODELS ----
     make_deep_ar()
     make_nbeats()
+    make_gp_forecaster()
 
 
 }